Added a separate Search.java class so that other classes can access it and use it.
